🐛 Fix: 버셀에서 잡은 오류 해결(다이나믹 렌더링을 위한 Suspense fallback 적용)
